requirements.txt
openai==1.10.0
psycopg==3.1.14
psycopg2-binary==2.9.9
python-dotenv==1.0.0
tiktoken==0.5.2
python-dotenv==1.0.0
pgvector==0.2.4
llama-index
sqlalchemy[asyncio]
chainlit==1.0.200
langchain==0.1.4
langchain-community==0.0.16
langchain-openai==0.0.5
pgvector
asyncpg
fastparquet
google-cloud-bigquery==3.17.1
db-dtypes==1.2.0
tabulate==0.9.0
chromadb==0.4.22